+++ src/main/java/com/example/surveyapp/domain/surveyanswer/domain/repository/SurveyOptionsAnswerRepository.java
@@ -1,6 +1,6 @@
 package com.example.surveyapp.domain.surveyanswer.domain.repository;
 
-import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyOptionsAnswer;
+import com.example.surveyapp.domain.surveyanswer.domain.model.entity.SurveyOptionsAnswer;
 
 public interface SurveyOptionsAnswerRepository {
     Long countByQuestionIdAndNumber(Long questionId, Long number);
