+++ src/main/java/com/example/surveyapp/domain/surveyanswer/domain/repository/SurveyTextAnswerRepository.java
@@ -1,6 +1,6 @@
 package com.example.surveyapp.domain.surveyanswer.domain.repository;
 
-import com.example.surveyapp.domain.surveyanswer.domain.model.SurveyTextAnswer;
+import com.example.surveyapp.domain.surveyanswer.domain.model.entity.SurveyTextAnswer;
 
 public interface SurveyTextAnswerRepository {
     SurveyTextAnswer save(SurveyTextAnswer surveyTextAnswer);
